# lab4/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080/api
    description: Generated server url
paths:
  /users/me:
    get:
      tags:
        - User Profile
      summary: Отримати дані свого профілю
      operationId: getMyProfile
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      tags:
        - User Profile
      summary: Оновити дані свого профілю
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/UserResponse"
  /users/me/password:
    put:
      tags:
        - User Profile
      summary: Змінити свій пароль
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "200":
          description: OK
  /organizations/{organizationId}/members/{userId}/role:
    put:
      tags:
        - Memberships
      summary: Оновити роль користувача
      operationId: updateRole
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
        required: true
      responses:
        "200":
          description: OK
...
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Скидання пароля
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Оновлення access токена
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Запит на відновлення пароля
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
...
components:
  schemas:
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 50
          minLength: 0
        lastName:
          type: string
          maxLength: 50
          minLength: 0
        email:
          type: string
          format: email
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          maxLength: 2147483647
          minLength: 6
      required:
        - newPassword
        - oldPassword
    UpdateRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum:
            - OWNER
            - ADMIN
            - ANALYST
            - VIEWER
      required:
        - role
...
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
        newPassword:
          type: string
          maxLength: 2147483647
          minLength: 6
      required:
        - newPassword
        - token
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    RefreshTokenRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
      required:
        - token
...
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
      required:
        - email
...